+++ src/Settings.ts
@@ -5,22 +5,19 @@ class Settings {
     private static conf = 'simplyblame';
 
     static getHeatMapColors(): string[] {
-        return vscode.workspace.getConfiguration(this.conf).heatMapColors;
+        if ([vscode.ColorThemeKind.Dark, vscode.ColorThemeKind.HighContrast].includes(vscode.window.activeColorTheme.kind)) {
+            return vscode.workspace.getConfiguration(this.conf).heatMapColorsDark;
+        } else {
+            return vscode.workspace.getConfiguration(this.conf).heatMapColorsLight;
+        }
     }
 
     static isEnableOpenBlameEditor(): boolean {
         return vscode.workspace.getConfiguration(this.conf).enableOpenBlameEditor;
     }
 
-    static getDateLocale(): string {
-        const lang = vscode.workspace.getConfiguration(this.conf).dateLocale;
-        switch (lang) {
-            case 'english':
-                return 'en-EN';
-            case 'finnish':
-            default:
-                return 'fi-FI';
-        }
+    static getDateFormat(): string {
+        return vscode.workspace.getConfiguration(this.conf).dateFormat;
     }
 }